vit.py
from transformers import ViTImageProcessor, ViTModel
from PIL import Image
import requests

url = 'http://images.cocodataset.org/val2017/000000039769.jpg'
image = Image.open(requests.get(url, stream=True).raw)

processor = ViTImageProcessor.from_pretrained('google/vit-base-patch16-224-in21k')
model = ViTModel.from_pretrained('google/vit-base-patch16-224-in21k')
inputs = processor(images=image, return_tensors="pt")

outputs = model(**inputs)
last_hidden_states = outputs.last_hidden_state

# output is 1x197x768, goal is to get 1D vector
# for now, use global average pooling
pooled_output = last_hidden_states.mean(dim=1)
embedding = pooled_output.flatten()

print(embedding)
print(embedding.shape)